RSE-246: Change TabHeader class and fixed as per comment on PR
